Release v0.2.0: Control the pairs you keep in an interaction model
New feature
-----------

distance_filter is a new argument for the build_model function that allows the
user to pass a dictionary containing pair distances. Any pairs whose separation
distance is given (within the tolerance of the distance bins) will be kept, all
other pairs will be removed. If distance_filter is not specified, then the
standard behavior of including all pairs within the distance r is assumed.

Refactoring
-----------

Grouping and ranking the distance bins now happens as part of the
count_neighbors routine. This information is then joined to the main neighbor
counts data frame when building the model coefficients data frame with the
build_model function.

Some functions and attributes were renamed to be more consistent with
refactorings that I implemented in sPyns.

All code has been passed through the Black code formatter.